requirements.txt
asgiref==3.8.1
cramjam==2.8.3
Django==5.0.7
djangorestframework==3.15.2
dnspython==2.6.1
pymongo==4.8.0
python-snappy==0.7.2
sqlparse==0.5.1
tzdata==2024.1
winkerberos==0.11.0
